Split renderer + add tests
