+++ js/renderer.ts
@@ -1,34 +1,15 @@
 import {
-    Alignment,
     Collider,
     CollisionComponent,
     Component,
-    Font,
     Material,
-    Mesh,
-    Object3D,
-    TextComponent,
-    TextEffect,
-    Texture,
-    TextWrapMode,
-    VerticalAlignment,
     ViewComponent,
     WonderlandEngine,
 } from '@wonderlandengine/api';
 import {property} from '@wonderlandengine/api/decorators.js';
 import {mat4, vec3} from 'gl-matrix';
 import {ReactNode} from 'react';
-import type {
-    ValueType,
-    ValueTypeNoAuto,
-    Color,
-    YogaNodeProps,
-    TextProps,
-    RoundedRectangleProps,
-    MeshProps,
-    NineSliceProps,
-    ReactComp,
-} from './renderer-types.js';
+import type {ReactComp} from './renderer-types.js';
 
 // Re-export types so other modules that import from './renderer.js' keep working
 export type {
@@ -43,17 +24,6 @@ export type {
     ReactComp,
 } from './renderer-types.js';
 
-import Reconciler, {Fiber, HostConfig} from 'react-reconciler';
-import {
-    Yoga,
-    Config,
-    Node as YogaNode,
-    MeasureMode,
-    /* We want to defer initializing the web assembly until
-     * the renderer is initialized. The only way we found was
-     * to import the files directly */
-} from 'yoga-layout/load';
-
 /* These are used by props */
 export {
     Align,
@@ -66,738 +36,20 @@ export {
 } from 'yoga-layout/load';
 
 /* These are used by the renderer only */
-import {
-    Align,
-    Display,
-    Edge,
-    Direction,
-    FlexDirection,
-    Gutter,
-    Justify,
-    Overflow,
-    PositionType,
-    Wrap,
-    loadYoga,
-} from 'yoga-layout/load';
+import {Direction} from 'yoga-layout/load';
 
-import {roundedRectangle, roundedRectangleOutline} from './rounded-rectangle-mesh.js';
 import {Cursor, CursorTarget, EventTypes} from '@wonderlandengine/components';
-import {nineSlice} from './nine-slice.js';
-import {
-    computeTextScaleAndWrap,
-    computeDimensionsFromBoundingBox,
-    TEXT_BASE_SIZE,
-    DEFAULT_FONT_SIZE,
-} from './renderer/text-helpers.js';
-import {
-    Z_INC,
-    setPositionCenter,
-    setPositionLeft,
-    setPositionRight,
-} from './renderer/position-helpers.js';
-import {
-    buildRoundedRectangleMeshProps,
-    buildNineSliceProps,
-} from './renderer/mesh-helpers.js';
-import {computeMeshChildTransforms} from './renderer/mesh-transform-helpers.js';
-import {propsEqual} from './renderer/props-helpers.js';
 
 // TEXT_BASE_SIZE and DEFAULT_FONT_SIZE are provided by renderer-text-helpers
 
-let Y: Yoga | null = null;
-
-export function destroyTreeForNode(child: NodeWrapper, ctx: Context) {
-    const childCount = child.children.length ?? 0;
-    for (let c = childCount - 1; c >= 0; --c) {
-        destroyTreeForNode(child.children[c], ctx);
-    }
-
-    if (child.parent) {
-        const parent = child.parent;
-        parent.node.removeChild(child.node);
-        parent.children.splice(parent.children.indexOf(child)!, 1);
-        child.parent = null;
-    }
-    child.node.free();
-
-    if (child.object && !child.object.isDestroyed) {
-        child.object.destroy();
-    }
-    child.object = null;
-
-    child.ctx?.removeNodeWrapper(child);
-}
-
-// propsEqual is implemented in js/renderer-helpers.ts
-
-class NodeWrapper {
-    node: YogaNode;
-    tag: string;
-    /* Applied properties cache */
-    props: any = {};
-    object: Object3D | null = null;
-    parent: NodeWrapper | null = null;
-    children: NodeWrapper[] = [];
-    ctx: Context;
-    hovering = [false, false];
-    dirty = true;
-
-    constructor(ctx: Context, node: YogaNode, tag: string) {
-        this.ctx = ctx;
-        this.tag = tag;
-        this.node = node;
-    }
-}
-
-class Context {
-    root: NodeWrapper | null;
-    config: Config;
-    comp: ReactComp;
-    wrappers: NodeWrapper[] = [];
-    constructor(c: ReactComp) {
-        this.root = null;
-        this.comp = c;
-
-        this.config = Y!.Config.create();
-        this.config.setUseWebDefaults(false);
-        this.config.setPointScaleFactor(1);
-    }
-
-    addNodeWrapper(w: NodeWrapper) {
-        this.wrappers.push(w);
-        w.ctx = this;
-    }
-    removeNodeWrapper(w: NodeWrapper) {
-        const i = this.wrappers.indexOf(w);
-        this.wrappers.splice(i, 1);
-    }
-
-    printTree(node?: NodeWrapper, prefix?: string) {
-        node = node ?? this.root!;
-        prefix = prefix ?? '';
-
-        const yn = node.node;
-        debug(
-            prefix + node.tag,
-            `{${yn.getComputedLeft()}, ${yn.getComputedTop()}, ${yn.getComputedWidth()}, ${yn.getComputedHeight()}}`,
-            node.props
-        );
-        if (!node.children) return;
-        for (let n of node.children) {
-            this.printTree(n, prefix + '--');
-        }
-    }
-}
-
-// position setters implemented in js/renderer-helpers.ts
-
-const tempBBVec4 = new Float32Array(4);
-function computeTextDimensions(
-    n: NodeWrapper,
-    context: Context
-): {width: number; height: number} {
-    const o = n.object!;
-    const t = o.getComponent(TextComponent)!;
-
-    const {scale, wrapWidth} = computeTextScaleAndWrap(
-        n.node.getComputedWidth(),
-        n.props.fontSize,
-        context.comp.scaling as [number, number],
-        TEXT_BASE_SIZE,
-        DEFAULT_FONT_SIZE
-    );
-
-    o.setScalingLocal([scale, scale, scale]);
-
-    if (!isNaN(wrapWidth)) {
-        t.wrapWidth = wrapWidth;
-        t.getBoundingBoxForText(n.props.text, tempBBVec4);
-        const {width, height} = computeDimensionsFromBoundingBox(
-            tempBBVec4,
-            scale,
-            context.comp.scaling as [number, number]
-        );
-        return {width, height};
-    }
-    return {width: 0, height: 0};
-}
-
-function applyLayoutToSceneGraph(n: NodeWrapper, context: Context, force?: boolean) {
-    debug('applyLayoutToSceneGraph');
-
-    if (!force && !n.dirty && !n.node.hasNewLayout()) return;
-    n.node.markLayoutSeen();
-
-    if (n.object?.isDestroyed) n.object = null;
-    const o =
-        n.object ??
-        (() => {
-            if (n.parent?.object?.isDestroyed) n.parent.object = null;
-            const o = context.comp.engine.scene.addObject(
-                n.parent?.object ?? context.comp.object
-            );
-            return o;
-        })();
-    n.object = o;
-    o.parent = n.parent?.object ?? context?.comp?.object;
-    o.resetPositionRotation();
-    o.resetScaling();
-
-    if (n.tag === 'text3d') {
-        const align = n.props.textAlign;
-        let alignment = Alignment.Left;
-        if (align === 'center') {
-            setPositionCenter(o, n, context.comp.scaling);
-            alignment = Alignment.Center;
-        } else if (align === 'right') {
-            setPositionRight(o, n, context.comp.scaling);
-            alignment = Alignment.Right;
-        } else {
-            setPositionLeft(o, n, context.comp.scaling);
-        }
-        const wrap = n.props.textWrap;
-        let textWrapMode = TextWrapMode.None;
-        if (wrap === 'soft') {
-            textWrapMode = TextWrapMode.Soft;
-        } else if (wrap === 'hard') {
-            textWrapMode = TextWrapMode.Hard;
-        } else if (wrap === 'clip') {
-            textWrapMode = TextWrapMode.Clip;
-        }
-
-        let t = o.getComponent(TextComponent);
-        if (!t) {
-            t = o.addComponent(TextComponent, {
-                text: n.props.text,
-                alignment,
-                effect: TextEffect.Outline,
-                verticalAlignment: VerticalAlignment.Top,
-                wrapMode: textWrapMode,
-                material: n.props.material ?? context.comp.textMaterial,
-            });
-        }
-        const s =
-            TEXT_BASE_SIZE *
-            (n.props.fontSize ?? DEFAULT_FONT_SIZE) *
-            context.comp.scaling[1];
-        o.setScalingLocal([s, s, s]);
-
-        t.material = n.props.material ?? context.comp.textMaterial;
-        const ww = (n.node.getComputedWidth() * context.comp.scaling[0]) / s;
-
-        if (!isNaN(ww)) {
-            t.wrapWidth = ww;
-        }
-    } else {
-        /* "mesh" and everything else */
-        if (context && context.comp && context.comp.scaling) {
-            setPositionLeft(o, n, context.comp.scaling);
-        }
-    }
-
-    if (n.tag === 'mesh' || n.tag === 'roundedRectangle' || n.tag === 'nineSlice') {
-        /* To offset the mesh, but avoid offsetting the children,
-         * we need to add a child object */
-        const child =
-            o.findByNameDirect('mesh')[0] ??
-            (() => {
-                const child = context.comp.engine.scene.addObject(o);
-                child.name = 'mesh';
-                return child;
-            })();
-
-        let sw = n.node.getComputedWidth() * context.comp.scaling[0];
-        let sh = n.node.getComputedHeight() * context.comp.scaling[1];
-
-        // there is a change that on the first time the code is executed getComputedWidth
-        // and getComputedHeight return NaN. In that case we set the values to 0 to prevent
-        // any more issues.
-        if (isNaN(sw) || isNaN(sh)) {
-            sw = 0;
-            sh = 0;
-        }
-        const transforms = computeMeshChildTransforms(sw, sh, n.props.z, n.tag);
-
-        const m = child.getComponent('mesh') ?? child.addComponent('mesh', {});
-        m.material = n.props.material;
-
-        let mesh = m.mesh;
-        if (n.tag === 'roundedRectangle') {
-            const p = buildRoundedRectangleMeshProps(
-                n.props,
-                sw,
-                sh,
-                context.comp.scaling as [number, number]
-            );
-            const props = (m as any).roundedRectangleProps ?? {};
-            const needsUpdate = !propsEqual(props, p);
-
-            const borderSize = (n.props.borderSize ?? 0) * context.comp.scaling[0];
-            if (needsUpdate) {
-                mesh = roundedRectangle(
-                    context.comp.engine,
-                    p.sw,
-                    p.sh,
-                    p.rounding,
-                    p.resolution,
-                    {
-                        tl: p.tl,
-                        tr: p.tr,
-                        bl: p.bl,
-                        br: p.br,
-                    },
-                    mesh
-                );
-                (m as any).roundedRectangleProps = p;
-            }
-
-            const oldBorderSize = (m as any).borderSize ?? 0;
-            const needsBorderMeshUpdate = oldBorderSize != borderSize;
-            const bm = child.getComponent('mesh', 1) ?? child.addComponent('mesh', {});
-            bm.active = borderSize != 0;
-            bm.material = n.props.borderMaterial;
-            if (needsBorderMeshUpdate) {
-                if (borderSize != 0) {
-                    bm.mesh = roundedRectangleOutline(
-                        context.comp.engine,
-                        p.sw,
-                        p.sh,
-                        p.rounding,
-                        p.resolution,
-                        borderSize,
-                        {
-                            tl: p.tl,
-                            tr: p.tr,
-                            bl: p.bl,
-                            br: p.br,
-                        },
-                        bm.mesh
-                    );
-                    (m as any).borderSize = borderSize;
-                }
-            }
-
-            child.setPositionLocal(transforms.position);
-            if (transforms.resetScaling) child.resetScaling();
-        } else if (n.tag === 'nineSlice') {
-            const p = buildNineSliceProps(
-                n.props,
-                sw,
-                sh,
-                context.comp.scaling as [number, number]
-            );
-            const props = (m as any).nineSliceProps ?? {};
-            const needsUpdate = !propsEqual(props, p);
-
-            if (needsUpdate) {
-                mesh = nineSlice(
-                    context.comp.engine,
-                    p.sw,
-                    p.sh,
-                    p.borderSize,
-                    p.borderTextureSize,
-                    mesh
-                );
-                (m as any).nineSliceProps = p;
-            }
-
-            child.setPositionLocal(transforms.position);
-            if (transforms.resetScaling) child.resetScaling();
-        } else {
-            /* Planes are diameter of 2 */
-            child.setPositionLocal(transforms.position);
-            if (transforms.scaling) child.setScalingLocal(transforms.scaling);
-        }
-        m.mesh = n.props.mesh ?? mesh;
-    }
-    /* For children created earlier  */
-    n.children?.forEach((c) => {
-        applyLayoutToSceneGraph(c, context, force);
-        if (c.object && !c.object.isDestroyed) {
-            c.object.parent = n.object;
-        }
-    });
-}
-
-function applyToYogaNode(
-    tag: string,
-    node: YogaNode,
-    props: YogaNodeProps | TextProps | RoundedRectangleProps | MeshProps | NineSliceProps,
-    wrapper: NodeWrapper,
-    ctx?: Context
-) {
-    if (tag === 'text3d') {
-        const p = props as TextProps;
-        let t = wrapper.object?.getComponent(TextComponent)!;
-        if (!t) {
-            /* Apply properties relevant to text component here */
-            wrapper.props.text = p.text;
-            wrapper.props.textAlign = p.textAlign;
-            wrapper.props.textWrap = p.textWrap;
-            wrapper.props.material = p.material;
-            applyLayoutToSceneGraph(wrapper, ctx!, true);
-            t = wrapper.object?.getComponent(TextComponent)!;
-        }
-        if (t.text !== p.text) {
-            t.text = p.text;
-            node.markDirty(); // trigger recalculating the layout.
-        }
-        node.setWidth(props.width);
-        node.setHeight(props.height);
-    } else {
-        if (ctx) {
-            applyLayoutToSceneGraph(wrapper, ctx, true);
-        } else {
-            debug('Context is undefined, skipping applyLayoutToSceneGraph');
-        }
-        node.setWidth(props.width);
-        node.setHeight(props.height);
-    }
-
-    /* Properties that allow undefined should be assigned to `undefined`,
-     *
-     * Properties that have a default value and do not allow `undefined` should be
-     * assigned the default value if props.value is `undefined`.
-     */
-    if (wrapper.props.alignContent !== props.alignContent)
-        node.setAlignContent(props.alignContent ?? Align.FlexStart);
-    if (wrapper.props.alignItems !== props.alignItems)
-        node.setAlignItems(props.alignItems ?? Align.Stretch);
-    if (wrapper.props.alignSelf !== props.alignSelf)
-        // TODO This default was not documented!
-        node.setAlignSelf(props.alignSelf ?? Align.FlexStart);
-    if (wrapper.props.aspectRatio !== props.aspectRatio)
-        node.setAspectRatio(props.aspectRatio);
-
-    if (wrapper.props.display !== props.display)
-        node.setDisplay(props.display ?? Display.Flex);
-
-    if (wrapper.props.flex !== props.flex) node.setFlex(props.flex);
-    if (wrapper.props.flexDirection !== props.flexDirection)
-        node.setFlexDirection(props.flexDirection ?? FlexDirection.Column);
-    if (wrapper.props.flexBasis !== props.flexBasis) node.setFlexBasis(props.flexBasis);
-    if (wrapper.props.flexGrow !== props.flexGrow) node.setFlexGrow(props.flexGrow);
-    if (wrapper.props.flexShrink !== props.flexShrink) node.setFlexShrink(props.flexShrink);
-    if (wrapper.props.flexWrap !== props.flexWrap)
-        node.setFlexWrap(props.flexWrap ?? Wrap.NoWrap);
-
-    if (wrapper.props.isReferenceBaseline !== props.isReferenceBaseline)
-        node.setIsReferenceBaseline(props.isReferenceBaseline ?? false);
-
-    if (wrapper.props.gap !== props.gap) node.setGap(Gutter.All, props.gap);
-    if (wrapper.props.rowGap !== props.rowGap) node.setGap(Gutter.Row, props.rowGap);
-    if (wrapper.props.columnGap !== props.columnGap)
-        node.setGap(Gutter.Column, props.columnGap);
-
-    if (wrapper.props.justifyContent !== props.justifyContent)
-        node.setJustifyContent(props.justifyContent ?? Justify.FlexStart);
-
-    if (wrapper.props.border !== props.border) node.setBorder(Edge.All, props.border);
-    if (wrapper.props.borderTop !== props.borderTop)
-        node.setBorder(Edge.Top, props.borderTop);
-    if (wrapper.props.borderBottom !== props.borderBottom)
-        node.setBorder(Edge.Bottom, props.borderBottom);
-    if (wrapper.props.borderLeft !== props.borderLeft)
-        node.setBorder(Edge.Left, props.borderLeft);
-    if (wrapper.props.borderRight !== props.borderRight)
-        node.setBorder(Edge.Right, props.borderRight);
-
-    if (wrapper.props.margin !== props.margin) node.setMargin(Edge.All, props.margin);
-    if (wrapper.props.marginTop !== props.marginTop)
-        node.setMargin(Edge.Top, props.marginTop);
-    if (wrapper.props.marginBottom !== props.marginBottom)
-        node.setMargin(Edge.Bottom, props.marginBottom);
-    if (wrapper.props.marginLeft !== props.marginLeft)
-        node.setMargin(Edge.Left, props.marginLeft);
-    if (wrapper.props.marginRight !== props.marginRight)
-        node.setMargin(Edge.Right, props.marginRight);
-
-    if (wrapper.props.maxHeight !== props.maxHeight) node.setMaxHeight(props.maxHeight);
-    if (wrapper.props.maxWidth !== props.maxWidth) node.setMaxWidth(props.maxWidth);
-    if (wrapper.props.minHeight !== props.minHeight) node.setMinHeight(props.minHeight);
-    if (wrapper.props.minWidth !== props.minWidth) node.setMinWidth(props.minWidth);
-    if (wrapper.props.overflow !== props.overflow)
-        node.setOverflow(props.overflow ?? Overflow.Hidden);
-
-    if (wrapper.props.padding !== props.padding) node.setPadding(Edge.All, props.padding);
-    if (wrapper.props.paddingTop !== props.paddingTop)
-        node.setPadding(Edge.Top, props.paddingTop);
-    if (wrapper.props.paddingBottom !== props.paddingBottom)
-        node.setPadding(Edge.Bottom, props.paddingBottom);
-    if (wrapper.props.paddingLeft !== props.paddingLeft)
-        node.setPadding(Edge.Left, props.paddingLeft);
-    if (wrapper.props.paddingRight !== props.paddingRight)
-        node.setPadding(Edge.Right, props.paddingRight);
-
-    if (wrapper.props.position !== props.position)
-        node.setPositionType(props.position ?? PositionType.Relative);
-    if (wrapper.props.top !== props.top) node.setPosition(Edge.Top, props.top);
-    if (wrapper.props.bottom !== props.bottom) node.setPosition(Edge.Bottom, props.bottom);
-    if (wrapper.props.left !== props.left) node.setPosition(Edge.Left, props.left);
-    if (wrapper.props.right !== props.right) node.setPosition(Edge.Right, props.right);
-
-    if (wrapper) wrapper.props = props;
-}
+import {NodeWrapper, Context} from './renderer/core.js';
+import {applyLayoutToSceneGraph, applyToYogaNode} from './renderer/layout.js';
 
 const DEBUG_RENDERER = false;
 const DEBUG_EVENTS = false;
 const debug = DEBUG_RENDERER ? console.log : () => {};
 
-const HostConfig: HostConfig<
-    string,
-    YogaNodeProps,
-    Context,
-    NodeWrapper,
-    void,
-    null,
-    null,
-    Object3D,
-    Context,
-    any,
-    any,
-    any,
-    any
-> = {
-    getRootHostContext(context: Context) {
-        return context;
-    },
-    getChildHostContext(parentHostContext: Context) {
-        return parentHostContext;
-    },
-    shouldSetTextContent(tag: string) {
-        return false;
-    },
-    createTextInstance(text: string, ctx: Context, hostContext: Context, node: ReactNode) {
-        debug('createTextInstance', text, ctx, hostContext, node);
-    },
-    createInstance(tag: string, props: YogaNodeProps, ctx: Context) {
-        debug('createInstance', tag, props, ctx);
-        const node = Y!.Node.create(ctx.config);
-        const w = new NodeWrapper(ctx, node, tag);
-        ctx.addNodeWrapper(w);
-
-        if (tag === 'text3d') {
-            node.setMeasureFunc(
-                (
-                    width: number,
-                    widthMode: MeasureMode,
-                    height: number,
-                    heightMode: MeasureMode
-                ): {
-                    width: number;
-                    height: number;
-                } => {
-                    const t = w.object?.getComponent(TextComponent);
-                    if (t) {
-                        const s =
-                            TEXT_BASE_SIZE *
-                            (w.props.fontSize ?? DEFAULT_FONT_SIZE) *
-                            ctx.comp.scaling[1];
-
-                        const ww = (width * ctx.comp.scaling[0]) / s;
-                        t.wrapWidth = ww;
-                        const bb = t.getBoundingBox();
-                        const bbWidth = ((bb[2] - bb[0]) * s) / ctx.comp.scaling[0];
-                        let bbHeight = ((bb[3] - bb[1]) * s) / ctx.comp.scaling[1];
-                        const font = (t.material as any).getFont() as Font;
-
-                        if (font) {
-                            let h = 0;
-                            h = (font.emHeight * s) / ctx.comp.scaling[1];
-                            // if the text is a single line,
-                            // we need to make sure the height is constant
-                            // because the bounding box is often a bit bigger a bit extra is added
-                            if (
-                                h > bbHeight ||
-                                t.wrapMode === TextWrapMode.None ||
-                                t.wrapMode === TextWrapMode.Clip
-                            ) {
-                                bbHeight = h;
-                            }
-                        }
-
-                        let calulatedWidth = 0;
-                        if (widthMode === MeasureMode.Undefined) {
-                            calulatedWidth = bbWidth;
-                        } else if (widthMode === MeasureMode.Exactly) {
-                            calulatedWidth = width;
-                        } else if (widthMode === MeasureMode.AtMost) {
-                            calulatedWidth = Math.min(bbWidth, width);
-                        }
-
-                        let calulatedHeight = 0;
-                        if (heightMode === MeasureMode.Undefined) {
-                            calulatedHeight = bbHeight;
-                        } else if (heightMode === MeasureMode.Exactly) {
-                            calulatedHeight = height;
-                        } else if (heightMode === MeasureMode.AtMost) {
-                            calulatedHeight = Math.min(bbHeight, height);
-                        }
-
-                        node.setHeight(calulatedHeight);
-                        node.setWidth(calulatedWidth);
-                        return {width: calulatedWidth, height: calulatedHeight};
-                    } else {
-                        return {width: width, height: height};
-                    }
-                }
-            );
-        }
-
-        applyToYogaNode(tag, node, props, w, ctx);
-        return w;
-    },
-    appendInitialChild(parent: NodeWrapper, child?: NodeWrapper) {
-        debug('appendInitialChild', child, parent);
-        if (!child) return;
-
-        applyToYogaNode(child.tag, child.node, child.props, child);
-        parent.node.insertChild(child.node, parent.node.getChildCount());
-
-        child.parent = parent;
-        parent.children!.push(child);
-
-        parent.ctx!.comp.needsUpdate = true;
-    },
-    appendChild(parent: NodeWrapper, child?: NodeWrapper) {
-        debug('appendChild', parent, child);
-        if (!child) return;
-
-        applyToYogaNode(child.tag, child.node, child.props, child);
-        parent.node.insertChild(child.node, parent.node.getChildCount());
-
-        child.parent = parent;
-        parent.children!.push(child);
-
-        parent.ctx!.comp.needsUpdate = true;
-    },
-    appendChildToContainer(ctx: Context, child?: NodeWrapper) {
-        debug('appendChildToContainer', ctx, child);
-        if (!child) return;
-        ctx.root = child;
-
-        ctx.comp.needsUpdate = true;
-    },
-
-    insertInContainerBefore(container, child: NodeWrapper | undefined, beforeChild) {
-        debug('insertContainerBefore', parent, child, beforeChild);
-        if (child === undefined) return;
-    },
-    insertBefore(parent: NodeWrapper, child: NodeWrapper | undefined, before: NodeWrapper) {
-        debug('insertBefore', parent, child, before);
-        if (!child) return;
-
-        applyToYogaNode(child.tag, child.node, child.props, child);
-
-        // Find the index of the 'before' node to determine the position at which to insert the new node
-        const beforeIndex = parent.children.findIndex((childNode) => childNode === before);
-        if (beforeIndex !== -1) {
-            parent.node.insertChild(child.node, beforeIndex);
-        }
-
-        child.parent = parent;
-
-        // We also need to insert the child in the correct position in the parent's children array
-        if (beforeIndex !== -1) {
-            parent.children.splice(beforeIndex, 0, child);
-        } else {
-            // If the 'before' node is not found, we will append the child by default.
-            parent.children.push(child);
-        }
-
-        parent.ctx!.comp.needsUpdate = true;
-    },
-    removeChild(parent: NodeWrapper, child?: NodeWrapper) {
-        debug('removeChild', parent, child);
-        if (!child) return;
-        destroyTreeForNode(child, parent.ctx!);
-
-        parent.ctx!.comp.needsUpdate = true;
-    },
-    removeChildFromContainer(ctx: Context, child?: NodeWrapper) {
-        debug('removeChildFromContainer', ctx, child);
-        if (!child) return;
-        destroyTreeForNode(child, ctx);
-    },
-    finalizeInitialChildren(
-        instance: NodeWrapper,
-        tag: string,
-        props: YogaNodeProps,
-        ctx: Context,
-        hostContext: Context
-    ) {
-        debug('finalizeInitialChildren', instance, tag);
-        return false;
-    },
-    prepareForCommit(ctx: Context) {
-        debug('prepareForCommit');
-        return null;
-    },
-    resetAfterCommit(containerInfo: Context) {
-        debug('resetAfterCommit', containerInfo);
-    },
-    commitUpdate(
-        instance: NodeWrapper,
-        updatePayload: null,
-        type: string,
-        prevProps: YogaNodeProps,
-        nextProps: YogaNodeProps,
-        internalHandle
-    ) {
-        debug('commitUpdate');
-        instance.props = nextProps;
-        instance.dirty = true;
-        applyToYogaNode(instance.tag, instance.node, instance.props, instance);
-
-        instance.ctx!.comp.needsUpdate = true;
-    },
-    prepareUpdate(
-        instance: NodeWrapper,
-        tag: string,
-        oldProps: YogaNodeProps,
-        newProps: YogaNodeProps,
-        rootContainer: Context,
-        hostContext: Context
-    ) {
-        debug('prepareUpdate', oldProps, newProps);
-        if (propsEqual(oldProps, newProps)) return null;
-        return {};
-    },
-    getPublicInstance(instance: NodeWrapper) {
-        return instance.object!;
-    },
-    afterActiveInstanceBlur() {},
-    beforeActiveInstanceBlur() {},
-    detachDeletedInstance(node: any) {},
-    getCurrentEventPriority() {
-        return 0;
-    },
-    getInstanceFromNode(node: any) {
-        return null;
-    },
-    clearContainer(ctx) {
-        debug('clearContainer', ctx);
-        if (!ctx.root) return;
-        destroyTreeForNode(ctx.root, ctx);
-        ctx.root = null;
-    },
-    getInstanceFromScope(scopeInstance: any) {
-        debug('getInstanceFromScope', scopeInstance);
-        return null;
-    },
-    scheduleTimeout: setTimeout,
-    cancelTimeout: clearTimeout,
-    supportsMutation: true,
-    supportsHydration: false,
-    supportsPersistence: false,
-    isPrimaryRenderer: false,
-    noTimeout: -1,
-    preparePortalMount(containerInfo: any) {},
-    prepareScopeUpdate(scopeInstance: any, instance: any) {},
-};
-const reconcilerInstance = Reconciler(HostConfig);
-
-// ReactComp type is exported from renderer-types.ts
+import {reconcilerInstance, initializeRenderer} from './renderer/reconciler-host.js';
 
 export enum UISpace {
     World = 0,
@@ -1219,61 +471,4 @@ export abstract class ReactUiBase extends Component implements ReactComp {
 
     abstract render(): ReactNode;
 }
-let yogaInitializationPromise: Promise<void> | null = null;
-
-type yogaRenderer = {
-    rootContainer: Reconciler.OpaqueRoot;
-    unmountRoot(): void;
-    render(element: ReactNode, reactComp: ReactComp, callback?: () => void): void;
-};
-
-export async function initializeRenderer() {
-    if (!Y) {
-        if (!yogaInitializationPromise) {
-            yogaInitializationPromise = loadYoga().then((loadedYoga) => {
-                Y = loadedYoga;
-            });
-        }
-
-        await yogaInitializationPromise;
-    }
-    return {
-        rootContainer: null,
-        unmountRoot() {
-            reconcilerInstance.updateContainer(null, this.rootContainer);
-        },
-        render(element: ReactNode, reactComp: ReactComp, callback?: () => void) {
-            const container = reconcilerInstance.createContainer(
-                new Context(reactComp),
-                0,
-                null,
-                false,
-                null,
-                'root',
-                (e: Error) => console.error(e),
-                null
-            );
-            this.rootContainer = container;
-            reactComp.setContext(container.containerInfo);
-
-            const parentComponent = null;
-            reconcilerInstance.updateContainer(
-                element,
-                container,
-                parentComponent,
-                reactComp.renderCallback.bind(reactComp)
-            );
-            const result = reconcilerInstance.injectIntoDevTools({
-                bundleType: 0,
-                version: '0.2.1',
-                rendererPackageName: 'wonderlandengine/react-ui',
-                //Reconciler.DevToolsConfig<NodeWrapper, void, any>
-                findFiberByHostInstance(instance: NodeWrapper | void): Fiber | null {
-                    console.log(instance ?? 'no instance');
-                    return null;
-                },
-            });
-            console.log('Devtools:', result);
-        },
-    };
-}
+// Reconciler initialization and renderer are provided by './renderer/reconciler-host.js'
